modify to override 1.5.0.17

+++ classes/stock/StockAvailable.php
@@ -20,7 +20,7 @@
 *
 *  @author PrestaShop SA <contact@prestashop.com>
 *  @copyright  2007-2012 PrestaShop SA
-*  @version  Release: $Revision: 17258 $
+*  @version  Release: $Revision: 17357 $
 *  @license    http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
 *  International Registered Trademark & Property of PrestaShop SA
 */
@@ -95,6 +95,9 @@ class StockAvailableCore extends ObjectModel
 	 */
 	public static function getStockAvailableIdByProductId($id_product, $id_product_attribute = null, $id_shop = null)
 	{
+		if (!Validate::isUnsignedId($id_product))
+			return false;
+
 		$query = new DbQuery();
 		$query->select('id_stock_available');
 		$query->from('stock_available');
@@ -114,6 +117,9 @@ class StockAvailableCore extends ObjectModel
 	 */
 	public static function synchronize($id_product, $order_id_shop = null)
 	{
+		if (!Validate::isUnsignedId($id_product))
+			return false;
+
 		// gets warehouse ids grouped by shops
 		$ids_warehouse = Warehouse::getWarehousesGroupedByShops();
 		if ($order_id_shop !== null)
@@ -240,6 +246,8 @@ class StockAvailableCore extends ObjectModel
 	 */
 	public static function setProductDependsOnStock($id_product, $depends_on_stock = true, $id_shop = null, $id_product_attribute = 0)
 	{
+		if (!Validate::isUnsignedId($id_product))
+			return false;
 
 		$existing_id = StockAvailable::getStockAvailableIdByProductId((int)$id_product, (int)$id_product_attribute, $id_shop);
 		if ($existing_id > 0)
@@ -275,6 +283,9 @@ class StockAvailableCore extends ObjectModel
 	 */
 	public static function setProductOutOfStock($id_product, $out_of_stock = false, $id_shop = null, $id_product_attribute = 0)
 	{
+		if (!Validate::isUnsignedId($id_product))
+			return false;
+
 		if ($id_shop === null)
 			$id_shop = Context::getContext()->shop->id;
 
@@ -388,6 +399,9 @@ class StockAvailableCore extends ObjectModel
 	 */
 	public static function updateQuantity($id_product, $id_product_attribute, $delta_quantity, $id_shop = null)
 	{
+		if (!Validate::isUnsignedId($id_product))
+			return false;
+
 		$id_stock_available = StockAvailable::getStockAvailableIdByProductId($id_product, $id_product_attribute, $id_shop);
 
 		if (!$id_stock_available)
@@ -428,6 +442,9 @@ class StockAvailableCore extends ObjectModel
 	 */
 	public static function setQuantity($id_product, $id_product_attribute, $quantity, $id_shop = null)
 	{
+		if (!Validate::isUnsignedId($id_product))
+			return false;
+
 		$context = Context::getContext();
 
 		// if there is no $id_shop, gets the context one
@@ -440,7 +457,6 @@ class StockAvailableCore extends ObjectModel
 		if (!$depends_on_stock)
 		{
 			$id_stock_available = (int)StockAvailable::getStockAvailableIdByProductId($id_product, $id_product_attribute, $id_shop);
-
 			if ($id_stock_available)
 			{
 				$stock_available = new StockAvailable($id_stock_available);
@@ -456,24 +472,19 @@ class StockAvailableCore extends ObjectModel
 				$stock_available->id_product_attribute = (int)$id_product_attribute;
 				$stock_available->quantity = (int)$quantity;
 
-				// if we are in shop_group context
-				if (Shop::getContext() == Shop::CONTEXT_GROUP)
+				$shop_group = new ShopGroup((int)Shop::getContextShopGroupID());
+		
+				// if quantities are shared between shops of the group
+				if ($shop_group->share_stock)
 				{
-					$shop_group = new ShopGroup((int)Shop::getContextShopGroupID());
-
-					// if quantities are shared between shops of the group
-					if ($shop_group->share_stock)
-					{
-						$stock_available->id_shop = 0;
-						$stock_available->id_shop_group = (int)$shop_group->id;
-					}
+					$stock_available->id_shop = 0;
+					$stock_available->id_shop_group = (int)$shop_group->id;
 				}
 				else
 				{
 					$stock_available->id_shop = $id_shop;
 					$stock_available->id_shop_group = Shop::getGroupFromShop($id_shop);
 				}
-
 				$stock_available->add();
 			}
 
@@ -496,12 +507,13 @@ class StockAvailableCore extends ObjectModel
 	 */
 	public static function removeProductFromStockAvailable($id_product, $id_product_attribute = null, $shop = null)
 	{
+		if (!Validate::isUnsignedId($id_product))
+			return false;
 		return Db::getInstance()->execute('
-			DELETE FROM '._DB_PREFIX_.'stock_available
-			WHERE id_product = '.(int)$id_product.
-			($id_product_attribute ? ' AND id_product_attribute = '.(int)$id_product_attribute : '').
-			StockAvailable::addSqlShopRestriction(null, $shop)
-		);
+		DELETE FROM '._DB_PREFIX_.'stock_available
+		WHERE id_product = '.(int)$id_product.
+		($id_product_attribute ? ' AND id_product_attribute = '.(int)$id_product_attribute : '').
+		StockAvailable::addSqlShopRestriction(null, $shop));
 	}
 
 	/**
@@ -545,6 +557,9 @@ class StockAvailableCore extends ObjectModel
 	 */
 	public static function dependsOnStock($id_product, $id_shop = null)
 	{
+		if (!Validate::isUnsignedId($id_product))
+			return false;
+
 		$query = new DbQuery();
 		$query->select('depends_on_stock');
 		$query->from('stock_available');
@@ -565,6 +580,9 @@ class StockAvailableCore extends ObjectModel
 	 */
 	public static function outOfStock($id_product, $id_shop = null)
 	{
+		if (!Validate::isUnsignedId($id_product))
+			return false;
+
 		$query = new DbQuery();
 		$query->select('out_of_stock');
 		$query->from('stock_available');
